Completo

+++ GerenciadorDeRede.sh
@@ -1,3 +1,4 @@
+
 #!/bin/bash
 
 ###########################################################
@@ -22,28 +23,28 @@ REDE=$( dialog					       \
 	5 "Desligar rede"		               \
 	6 "Ligar rede"			               \
 	7 "Reiniciar redes"             	       \
-	8 "Remover todos os endereços IPs de uma rede"  \
+	8 "Remover todos os endereços IPs de uma rede" \
 	9 "Mostrar a configuração de um IP"            \
 	10 "Mostrar todos IPs configurados"            \
 	11 "Configurar uma rede via DHCP"              \
-	12 "Pingar um IP configurado"                 \
-	13 "Voltar")
+	12 "Pingar um IP configurado"                  \
+	13 "Voltar"				       )
 
 case $REDE in
-1) CFUR ;;
-2) CFDR ;;
-3) CFTR ;;
-4) CFQR ;;
-5) DSRE ;;
-6) LGRE ;;
-7) RERR ;;
-8) RMTP ;;
-9) MCIP ;;
-10) MTIP ;;
-11) CURD ;;
-12) PICO ;;
-13) ./menu.sh ;;
-*) PRESSIONE;;
+1) CFUR 			    ;;
+2) CFDR 			    ;;
+3) CFTR 			    ;;
+4) CFQR		 		    ;;
+5) DSRE 			    ;;
+6) LGRE 			    ;;
+7) RERR 			    ;;
+8) RMTP 			    ;;
+9) MCIP 			    ;;
+10) MTIP 			    ;;
+11) CURD                            ;;
+12) PICO 			    ;;
+13) /home/vagrant/Projeto/./menu.sh ;;
+*) PRESSIONE			    ;;
 esac
 }
 PRESSIONE(){
@@ -53,35 +54,42 @@ clear
 }
 # Função responsável pela configuração de uma rede.
 CFUR(){
-IP=$( dialog --stdout --inputbox 'Digite o número IP do eth0: ' 0 0 )
-MASCARA=$( dialog --stdout --inputbox 'Digite a Máscara de Rede do eth0: ' 0 0 )
-NETWORK=$( dialog --stdout --inputbox 'Digite a Network do eth0: ' 0 0 )
-BROADCAST=$( dialog --stdout --inputbox 'Digite o Broadcast do eth0: ' 0 0 )
+dialog --title "Aviso" --yesno "Você Quer Mesmo Continuar ?" 0 0
+if [ $? = 0 ]; then
+IP=$( dialog --stdout --nocancel --inputbox 'Digite o número IP do eth0: ' 0 0 )
+MASCARA=$( dialog --stdout --nocancel --inputbox 'Digite a Máscara de Rede do eth0: ' 0 0 )
+NETWORK=$( dialog --stdout --nocancel --inputbox 'Digite a Network do eth0: ' 0 0 )
+BROADCAST=$( dialog --stdout --nocancel --inputbox 'Digite o Broadcast do eth0: ' 0 0 )
 echo "   auto lo
    iface lo inet loopback
 
-   # Primeira Interface de Rede
-   auto eth0
-   allow-hotplug eth0
-   iface eth0 inet static
-   address $IP
-   netmask $MASCARA
-   network $NETWORK
-   broadcast $BROADCAST" > /etc/network/interfaces
+  # Primeira Interface de Rede
+  auto eth0
+  allow-hotplug eth0
+  iface eth0 inet static
+  address $IP
+  netmask $MASCARA
+  network $NETWORK
+  broadcast $BROADCAST" > /etc/network/interfaces
 dialog --stdout --msgbox "Rede configurada com sucesso " 0 0
 MENU
+else
+MENU
+fi
 }
 
 # Função responsável pela configuração de duas rede.
 CFDR(){
-IP=$( dialog --stdout --inputbox 'Digite o número IP do eth0: ' 0 0 )
-MASCARA=$( dialog --stdout --inputbox 'Digite a Máscara de rede do eth0: ' 0 0 )
-NETWORK=$( dialog --stdout --inputbox 'Digite a Network do eth0: ' 0 0 )
-BROADCAST=$( dialog --stdout --inputbox 'Digite o Broadcast do eth0: ' 0 0 )
-IP2=$( dialog --stdout --inputbox 'Digite o número IP do eth1: ' 0 0 )
-MASCARA2=$( dialog --stdout --inputbox 'Digite a Máscara de rede do eth1: ' 0 0 )
-NETWORK2=$( dialog --stdout --inputbox 'Digite a Network do eth1: ' 0 0 )
-BROADCAST2=$( dialog --stdout --inputbox 'Digite o Broadcast do eth1: ' 0 0 )
+dialog --title "Aviso" --yesno "Você Quer Mesmo Continuar ?" 0 0
+if [ $? = 0 ]; then
+IP=$( dialog --stdout --nocancel --inputbox 'Digite o número IP do eth0: ' 0 0 )
+MASCARA=$( dialog --stdout --nocancel --inputbox 'Digite a Máscara de rede do eth0: ' 0 0 )
+NETWORK=$( dialog --stdout --nocancel --inputbox 'Digite a Network do eth0: ' 0 0 )
+BROADCAST=$( dialog --stdout --nocancel --inputbox 'Digite o Broadcast do eth0: ' 0 0 )
+IP2=$( dialog --stdout --nocancel --inputbox 'Digite o número IP do eth1: ' 0 0 )
+MASCARA2=$( dialog --stdout --nocancel --inputbox 'Digite a Máscara de rede do eth1: ' 0 0 )
+NETWORK2=$( dialog --stdout --nocancel --inputbox 'Digite a Network do eth1: ' 0 0 )
+BROADCAST2=$( dialog --stdout -nocancel --inputbox 'Digite o Broadcast do eth1: ' 0 0 )
 echo "   auto lo
    iface lo inet loopback
 
@@ -103,22 +111,27 @@ echo "   auto lo
    broadcast $BROADCAST2" > /etc/network/interfaces
 dialog --msgbox "Redes configuradas com sucesso " 0 0
 MENU
+else
+MENU
+fi
 }
 
 # Função responsável pela configuração de três rede.
 CFTR(){
-IP=$( dialog --stdout --inputbox 'Digite o número IP do eth0: ' 0 0 )
-MASCARA=$( dialog --stdout --inputbox 'Digite a Máscara de rede do eth0: ' 0 0 )
-NETWORK=$( dialog --stdout --inputbox 'Digite a Network do eth0: ' 0 0 )
-BROADCAST=$( dialog --stdout --inputbox 'Digite o Broadcast do eth0: ' 0 0 )
-IP2=$( dialog --stdout --inputbox 'Digite o número IP do eth1: ' 0 0 )
-MASCARA2=$( dialog --stdout --inputbox 'Digite a Máscara de rede do eth1: ' 0 0 )
-NETWORK2=$( dialog --stdout --inputbox 'Digite a Network do eth1: ' 0 0 )
-BROADCAST2=$( dialog --stdout --inputbox 'Digite o Broadcast do eth1: ' 0 0 )
-IP3=$( dialog --stdout --inputbox 'Digite o número IP do eth2: ' 0 0 )
-MASCARA3=$( dialog --stdout --inputbox 'Digite a Máscara de rede do eth2: ' 0 0 )
-NETWORK3=$( dialog --stdout --inputbox 'Digite a Network do eth2: ' 0 0 )
-BROADCAST3=$( dialog --stdout --inputbox 'Digite o Broadcast do eth2: ' 0 0 )
+dialog --title "Aviso" --yesno "Você Quer Mesmo Continuar ?" 0 0
+if [ $? = 0 ]; then
+IP=$( dialog --stdout --nocancel --inputbox 'Digite o número IP do eth0: ' 0 0 )
+MASCARA=$( dialog --stdout --nocancel --inputbox 'Digite a Máscara de rede do eth0: ' 0 0 )
+NETWORK=$( dialog --stdout --nocancel --inputbox 'Digite a Network do eth0: ' 0 0 )
+BROADCAST=$( dialog --stdout --nocancel --inputbox 'Digite o Broadcast do eth0: ' 0 0 )
+IP2=$( dialog --stdout --nocancel --inputbox 'Digite o número IP do eth1: ' 0 0 )
+MASCARA2=$( dialog --stdout --nocancel --inputbox 'Digite a Máscara de rede do eth1: ' 0 0 )
+NETWORK2=$( dialog --stdout --nocancel --inputbox 'Digite a Network do eth1: ' 0 0 )
+BROADCAST2=$( dialog --stdout --nocancel --inputbox 'Digite o Broadcast do eth1: ' 0 0 )
+IP3=$( dialog --stdout --nocancel --inputbox 'Digite o número IP do eth2: ' 0 0 )
+MASCARA3=$( dialog --stdout --nocancel --inputbox 'Digite a Máscara de rede do eth2: ' 0 0 )
+NETWORK3=$( dialog --stdout --nocancel --inputbox 'Digite a Network do eth2: ' 0 0 )
+BROADCAST3=$( dialog --stdout --nocancel --inputbox 'Digite o Broadcast do eth2: ' 0 0 )
 echo "   auto lo
    iface lo inet loopback
 
@@ -148,26 +161,31 @@ echo "   auto lo
    broadcast $BROADCAST3" > /etc/network/interfaces
 dialog --msgbox "Redes configuradas com sucesso " 0 0
 MENU
+else
+MENU
+fi
 }
 
 # Função responsável pela configuração de quatro rede.
 CFQR(){
-IP=$( dialog --stdout --inputbox 'Digite o número IP do eth0: ' 0 0 )
-MASCARA=$( dialog --stdout --inputbox 'Digite a Máscara de rede do eth0: ' 0 0 )
-NETWORK=$( dialog --stdout --inputbox 'Digite a Network do eth0: ' 0 0 )
-BROADCAST=$( dialog --stdout --inputbox 'Digite o Broadcast do eth0: ' 0 0 )
-IP2=$( dialog --stdout --inputbox 'Digite o número IP do eth1: ' 0 0 )
-MASCARA2=$( dialog --stdout --inputbox 'Digite a Máscara de rede do eth1: ' 0 0 )
-NETWORK2=$( dialog --stdout --inputbox 'Digite a Network do eth1: ' 0 0 )
-BROADCAST2=$( dialog --stdout --inputbox 'Digite o Broadcast do eth1: ' 0 0 )
-IP3=$( dialog --stdout --inputbox 'Digite o número IP do eth2: ' 0 0 )
-MASCARA3=$( dialog --stdout --inputbox 'Digite a Máscara de rede do eth2: ' 0 0 )
-NETWORK3=$( dialog --stdout --inputbox 'Digite a Network do eth2: ' 0 0 )
-BROADCAST3=$( dialog --stdout --inputbox 'Digite o Broadcast do eth2: ' 0 0 )
-IP4=$( dialog --stdout --inputbox 'Digite o número IP do eth3: ' 0 0 )
-MASCARA4=$( dialog --stdout --inputbox 'Digite a Máscara de Rede do eth3: ' 0 0 )
-NETWORK4=$( dialog --stdout --inputbox 'Digite a Network do eth3: ' 0 0 )
-BROADCAST4=$( dialog --stdout --inputbox 'Digite o Broadcast do eth3: ' 0 0 )
+dialog --title "Aviso" --yesno "Você Quer Mesmo Continuar ?" 0 0
+if [ $? = 0 ]; then
+IP=$( dialog --stdout --nocancel --inputbox 'Digite o número IP do eth0: ' 0 0 )
+MASCARA=$( dialog --stdout --nocancel --inputbox 'Digite a Máscara de rede do eth0: ' 0 0 )
+NETWORK=$( dialog --stdout --nocancel --inputbox 'Digite a Network do eth0: ' 0 0 )
+BROADCAST=$( dialog --stdout --nocancel --inputbox 'Digite o Broadcast do eth0: ' 0 0 )
+IP2=$( dialog --stdout --nocancel --inputbox 'Digite o número IP do eth1: ' 0 0 )
+MASCARA2=$( dialog --stdout --nocancel --inputbox 'Digite a Máscara de rede do eth1: ' 0 0 )
+NETWORK2=$( dialog --stdout --nocancel --inputbox 'Digite a Network do eth1: ' 0 0 )
+BROADCAST2=$( dialog --stdout --nocancel --inputbox 'Digite o Broadcast do eth1: ' 0 0 )
+IP3=$( dialog --stdout --nocancel --inputbox 'Digite o número IP do eth2: ' 0 0 )
+MASCARA3=$( dialog --stdout --nocancel --inputbox 'Digite a Máscara de rede do eth2: ' 0 0 )
+NETWORK3=$( dialog --stdout --nocancel --inputbox 'Digite a Network do eth2: ' 0 0 )
+BROADCAST3=$( dialog --stdout --nocancel --inputbox 'Digite o Broadcast do eth2: ' 0 0 )
+IP4=$( dialog --stdout --nocancel --inputbox 'Digite o número IP do eth3: ' 0 0 )
+MASCARA4=$( dialog --stdout --nocancel --inputbox 'Digite a Máscara de Rede do eth3: ' 0 0 )
+NETWORK4=$( dialog --stdout --nocancel --inputbox 'Digite a Network do eth3: ' 0 0 )
+BROADCAST4=$( dialog --stdout --nocancel --inputbox 'Digite o Broadcast do eth3: ' 0 0 )
 echo "   auto lo
    iface lo inet loopback
 
@@ -206,11 +224,16 @@ echo "   auto lo
 
 dialog --msgbox "Redes configuradas com sucesso " 0 0
 MENU
+else
+MENU
+fi
 }
 
 # Função responsável por desligar a rede.
 DSRE(){
-REDE=$( dialog --stdout --inputbox 'Qual rede você deseja desligar ? ' 0 0 )
+dialog --title "Aviso" --yesno "Você Quer Mesmo Continuar ?" 0 0
+if [ $? = 0 ]; then
+REDE=$( dialog --stdout --nocancel --inputbox 'Qual rede você deseja desligar ? ' 0 0 )
 ifdown $REDE > REDE &
 echo '10' | dialog --stdout --gauge 'Desligando aguarde....' 0 0 0
 sleep 2
@@ -223,23 +246,31 @@ sleep 2
 dialog --stdout --msgbox 'Rede desligada com sucesso' 0 0
 rm REDE
 MENU
+else
+MENU
+fi
 }
 
 # Função responsável por ligar a rede.
 LGRE(){
-REDE=$( dialog --stdout --inputbox 'Qual rede você deseja ligar ? ' 0 0 )
+dialog --title "Aviso" --yesno "Você Quer Mesmo Continuar ?" 0 0
+if [ $? = 0 ]; then
+REDE=$( dialog --stdout --nocancel --inputbox 'Qual rede você deseja ligar ? ' 0 0 )
 ifup $REDE > REDE &
-echo '10' | dialog --stdout --gauge 'Ligando aguarde....' 0 0 0
+echo '10' | dialog --stdout --nocancel --gauge 'Ligando aguarde....' 0 0 0
 sleep 2
-echo '30' | dialog --stdout --gauge 'Ligando aguarde....' 0 0 0
+echo '30' | dialog --stdout --nocancel --gauge 'Ligando aguarde....' 0 0 0
 sleep 2
-echo '60' | dialog --stdout --gauge 'Ligando aguarde....' 0 0 0
+echo '60' | dialog --stdout --nocancel 'Ligando aguarde....' 0 0 0
 sleep 2
-echo '100' | dialog --stdout --gauge 'Ligando aguarde....' 0 0 0
+echo '100' | dialog --stdout --nocancel 'Ligando aguarde....' 0 0 0
 sleep 2
 dialog --stdout --msgbox 'Rede ligada com sucesso ' 0 0
 rm REDE
 MENU
+else
+MENU
+fi
 }
 
 # Função responsável por reiniciar a rede.
@@ -256,28 +287,39 @@ if [ $? = 0 ]; then
 	echo '100' | dialog --stdout --gauge 'Reiniciando aguarde....' 0 0 0
 	sleep 2
 	dialog --stdout --msgbox "Reiniciado com sucesso" 0 0
+	MENU
 else
 	dialog --msgbox "Operação cancelada" 0 0
+	MENU
 fi
-MENU
 }
 
 # Função responsável por remover todos os endereços IPs de uma rede.
 RMTP(){
-REDE=$( dialog --stdout --inputbox 'Qual rede você deseja remover os endereços IP ? ' 0 0 )
+dialog --title "Aviso" --yesno "Você Quer Mesmo Continuar ?" 0 0
+if [ $? = 0 ]; then
+REDE=$( dialog --stdout --nocancel --inputbox 'Qual rede você deseja remover os endereços IP ? ' 0 0 )
 ip addr flush dev $REDE > REDE
 dialog --msgbox 'Todos endereços IPs foram removidos com sucesso' 0 0
 rm REDE
 MENU
+else
+MENU
+fi
 }
 
 # Função responsável por mostrar a configuração de um ip.
 MCIP(){
+dialog --title "Aviso" --yesno "Você Quer Mesmo Continuar ?" 0 0
+if [ $? = 0 ]; then
 MOSTRAR=$( dialog --stdout --inputbox 'Digite a rede: ' 0 0 )
 ip addr show dev $MOSTRAR > MOSTRAR.txt
 dialog --textbox MOSTRAR.txt 0 0
 rm MOSTRAR.txt
 MENU
+else
+MENU
+fi
 }
 
 # Função responsável por mostrar todos os ip configurados.
@@ -290,19 +332,27 @@ MENU
 
 # Função responsável por configuração de uma rede por DHCP.
 CURD(){
-DHCP=$( dialog --stdout --inputbox 'Qual rede você deseja configurar ? ' 0 0 )
+dialog --title "Aviso" --yesno "Você Quer Mesmo Continuar ?" 0 0
+if [ $? = 0 ]; then
+DHCP=$( dialog --stdout --nocancel --inputbox 'Qual rede você deseja configurar ? ' 0 0 )
 dhclient $DHCP > DHCP
 dialog --msgbox  'Rede configurada com sucesso ' 0 0
 rm DHCP
 MENU
+fi
 }
 
 # Função responsável por pingar um ip configurado.
 PICO(){
-PINGAR=$( dialog --stdout --inputbox 'Qual o IP da rede ? ' 0 0 )
+dialog --title "Aviso" --yesno "Você Quer Mesmo Continuar ?" 0 0
+if [ $? = 0 ]; then
+PINGAR=$( dialog --stdout --nocancel --inputbox 'Qual o IP da rede ? ' 0 0 )
 ping -c 6 $PINGAR > PING &
 dialog --tailbox PING 70 70
 rm PING
 MENU
+else
+MENU
+fi
 }
 MENU
